Tweak pagure box so we can create projects
Previously, the way I set up the pagure box, we were running the main
pagure.service as root, primarily becasue we needed it to serve over
port 443. However, because of this the permissions were messed up after
creating a new project such that creating projects didnt wrok. Now we go
back to running the service under the git user (like is done in the
pagure upstream src vagrant box) and making copies of some of the files
pagure needs for OIDC auth to where pagure and the git user can see them

Signed-off-by: Ryan Lerch <rlerch@redhat.com>