WIP: Migrate to containers
This is a work in progress, and it does not currently work as the
containers aren't reachable from the host. I'm not finding a way to
define a podman network that would let the host reach the containers
using their IP and the original ports that the services running in the
containers listen to.

References:
- https://github.com/containers/podman/discussions/24823
- https://github.com/eriksjolund/podman-networking-docs
- https://github.com/containers/podman/discussions/22943#discussioncomment-9795883
- https://bugs.passt.top/show_bug.cgi?id=48

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>